Add logger for update function
