+++ lib/code_corps/github.ex
@@ -227,40 +227,4 @@ defmodule CodeCorps.GitHub do
   defp handle_response({:error, reason}) do
     {:error, HTTPClientError.new(reason: reason)}
   end
-
-  def create_issue(project, attributes, current_user) do
-    access_token = current_user.github_access_token || default_user_token() # need to create the Github user for this token
-    client = Tentacat.Client.new(%{access_token: access_token})
-    response = Tentacat.Issues.create(
-      project.github_owner,
-      project.github_repo,
-      attributes,
-      client
-    )
-    case response.status do
-      201 ->
-        response.body["id"] # return the github id
-      _ ->
-        Logger.error "Could not create task for Project ID: #{project.id}. Error: #{response.body}"
-    end
-  end
-
-  def update_issue(attributes, task, current_user) do
-    access_token = current_user.github_access_token || default_user_token() # need to create the Github user for this token
-    client = Tentacat.Client.new(%{access_token: access_token})
-    response = Tentacat.Issues.update(
-      task.project.github_owner,
-      task.project.github_repo,
-      task.github_id,
-      attributes,
-      client
-    )
-    unless response.status == 200 do
-      # log error
-    end
-  end
-
-  defp default_user_token do
-    System.get_env("GITHUB_DEFAULT_USER_TOKEN")
-  end
 end
