--- lib/code_corps/github_testing.ex
@@ -1,10 +1,0 @@
-defmodule CodeCorps.GithubTesting do
-  def create_issue(attributes, _project, _current_user) do
-    case attributes["error_testing"] do
-      true ->
-        nil
-      _ ->
-        1 # Return github id
-    end
-  end
-end
